+++ src/GamePack/Trail.java
@@ -8,12 +8,12 @@ public class Trail extends GameObject{
     private Handler handler;
     private Color color;
 
-    private int width , height;
+    private float width , height;
     private float life;
 
     // life=0.001 - 0.1
 
-    public Trail(int x, int y, ID id , Handler handler , Color color , int height, int width , float life) {
+    public Trail(float x, float y, ID id , Handler handler , Color color , float height, float width , float life) {
         super(x, y, id);
         this.handler = handler;
         this.color = color;
@@ -35,7 +35,7 @@ public class Trail extends GameObject{
         g2d.setComposite(makeTransparent(alpha));
 
         g.setColor(color);
-        g.fill3DRect((int)x,(int)y,width,height,true);
+        g.fill3DRect((int)x,(int)y,(int)width,(int)height,true);
 
         g2d.setComposite(makeTransparent(1));
